Removed unwanted files
