--- setup.py
@@ -1,9 +1,0 @@
-from distutils.core import setup
-
-setup(
-	name='Skin',
-	version='0.1dev',
-	packages=['skin',],
-	license='Open for Commercial as well as Non-Commercial Use'
-	# long_description=open('README.txt').read()
-)
